mosque_project/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-jazzmin==3.0.0
et-xmlfile==1.1.0
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.2
django-jazzmin-admin-rangefilter==1.0.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
